requirements.txt
pyserial
python-aqi
